requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
boto3==1.17.47
botocore==1.20.47
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
discord.py==1.7.3
dnspython==2.1.0
email-validator==1.1.2
google-auth==1.33.0
google-auth-oauthlib==0.4.4
gspread==3.7.0
httplib2==0.19.1
idna==2.10
iniconfig==1.1.1
jmespath==0.10.0
multidict==5.1.0
oauth2client==4.1.3
oauthlib==3.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.3
pytest-metadata==1.11.0
pytest-mock==3.5.1
pytest-timeout==1.4.2
python-dateutil==2.8.1
python-dotenv==0.18.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.3.6
six==1.15.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
yarl==1.6.3
youtube-dl==2021.6.6